+++ app/src/main/cpp/Game/Headers/frutils.h
@@ -21,23 +21,23 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #if (defined(powerc) || defined(__powerc))
 extern  void Fast_Slot_Copy(grs_bitmap *bm);
 extern  void Fast_FullScreen_Copy(grs_bitmap *bm);
-extern  void Fast_Slot_Double(grs_bitmap *bm, long w, long h);
-extern  void Fast_FullScreen_Double(grs_bitmap *bm, long w, long h);
+extern  void Fast_Slot_Double(grs_bitmap *bm, int32_t w, int32_t h);
+extern  void Fast_FullScreen_Double(grs_bitmap *bm, int32_t w, int32_t h);
 
-void FastSlotDouble2Canvas(grs_bitmap *bm, grs_canvas *destCanvas, long w, long h);
-void FastFullscreenDouble2Canvas(grs_bitmap *bm, grs_canvas *destCanvas, long w, long h);
+void FastSlotDouble2Canvas(grs_bitmap *bm, grs_canvas *destCanvas, int32_t w, int32_t h);
+void FastFullscreenDouble2Canvas(grs_bitmap *bm, grs_canvas *destCanvas, int32_t w, int32_t h);
 
 #else
 
 extern  asm void Fast_Slot_Copy(grs_bitmap *bm);
 extern  asm void Fast_FullScreen_Copy(grs_bitmap *bm);
-extern  asm void Fast_Slot_Double(grs_bitmap *bm, long w, long h);
-extern  asm void Fast_FullScreen_Double(grs_bitmap *bm, long w, long h);
+extern  asm void Fast_Slot_Double(grs_bitmap *bm, int32_t w, int32_t h);
+extern  asm void Fast_FullScreen_Double(grs_bitmap *bm, int32_t w, int32_t h);
 #endif
 
 
 // Stuff for the low-res temporary offscreen buffer.
 extern grs_canvas		gDoubleSizeOffCanvas;
 
-int AllocDoubleBuffer(int w, int h);
+int32_t AllocDoubleBuffer(int32_t w, int32_t h);
 void FreeDoubleBuffer(void);
