+++ app/src/main/cpp/Game/Headers/tilecam.h
@@ -42,7 +42,7 @@ typedef void (*camera_setfunc)(struct _tilecamera *tc);
 
 typedef struct _tilecamera
 {
-   long bcolor, fcolor;
+   int32_t bcolor, fcolor;
    fix x,y,theta;
    bool show;
    camera_setfunc func;
