+++ app/src/main/cpp/Game/cybrnd.c
@@ -54,8 +54,8 @@ RNDSTREAM_STD(start_rnd);
 
 void rnd_init(void)
 {
-   long  random_seed;
-   long  seed;
+   int32_t  random_seed;
+   int32_t  seed;
 
    // try to get a completely random value - take the bottom two bytes to be
    // the random seed
